Cleanup Precompiles (#2092)
* Move Celo preocompiles to celo_contracts.go (do the same with tests)
* Modify calling behaviour so that we don't need to modify every eth precompile (just use a wrapper for them)
* Simplify diff with contracts and contracts_test
