+++ tests/fuzzers/bls12381/precompile_fuzzer.go
@@ -18,7 +18,6 @@ package bls
 
 import (
 	"bytes"
-	"errors"
 	"fmt"
 	"math/big"
 
@@ -112,10 +111,6 @@ var vmBlockCtx = vm.BlockContext{
 	BlockNumber: new(big.Int).Set(testHeader.Number),
 	Time:        new(big.Int).SetUint64(testHeader.Time),
 
-	GetRegisteredAddress: func(evm *vm.EVM, registryId common.Hash) (common.Address, error) {
-		return common.ZeroAddress, errors.New("not implemented: GetAddressFromRegistry")
-	},
-
 	EpochSize:         100,
 	GetValidators:     getValidators,
 	GetHeaderByNumber: getHeaderByNumber,
@@ -186,7 +181,7 @@ func fuzz(id byte, data []byte) int {
 	}
 	cpy := make([]byte, len(data))
 	copy(cpy, data)
-	_, err := precompile.Run(cpy, common.HexToAddress("1337"), mockEVM)
+	_, err := precompile.Run(cpy, vm.NewContext(common.HexToAddress("1337"), mockEVM))
 	if !bytes.Equal(cpy, data) {
 		panic(fmt.Sprintf("input data modified, precompile %d: %x %x", id, data, cpy))
 	}
